requirements/base.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.7
Flask-JWT-Extended==4.6.0
Flask-CORS==5.0.0
Werkzeug==3.0.6
python-dotenv==1.0.1
cloudinary==1.41.0
boto3==1.35.24
pydantic==2.12.5
pydantic[email]==2.12.5
Flask-Limiter==3.8.0
Flask-Mail==0.10.0
Faker==39.0.0